+++ packages/gov-ui-kit/src/modules/components/index.ts
@@ -1,3 +1,4 @@
+export * from './address';
 export * from './dao';
 export * from './member';
 export * from './odsModulesProvider';