+++ packages/gov-ui-kit/src/modules/utils/addressUtils/addressUtils.test.tsx
@@ -1,0 +1,63 @@
+/**
+ * Set node as jest environment because the checksum check on the viem isAddress utility does not properly work on jsdom
+ * (See here https://github.com/jestjs/jest/issues/7780#issuecomment-865077151)
+ * @jest-environment node
+ */
+import { testLogger } from '../../../core/test';
+import { addressUtils } from './addressUtils';
+
+describe('address utils', () => {
+    describe('isAddress', () => {
+        it('returns true when the input is a valid address', () => {
+            const value = '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045';
+            expect(addressUtils.isAddress(value)).toBeTruthy();
+        });
+
+        it('returns false when the input is not a valid address', () => {
+            const value = 'name.eth';
+            expect(addressUtils.isAddress(value)).toBeFalsy();
+        });
+
+        it('returns true when the input is a valid address but not in the checksum format', () => {
+            const value = '0xd8dA6BF26964aF9D7eEd9e03E53415D37aa96045';
+            expect(addressUtils.isAddress(value)).toBeTruthy();
+        });
+
+        it('returns false when the input is a valid address but not in the checksum format when the strict property is set to true', () => {
+            const value = '0xe11bFCBDd43745d4Aa6f4f18E24aD24f4623af04';
+            const options = { strict: true };
+            expect(addressUtils.isAddress(value, options)).toBeFalsy();
+        });
+    });
+
+    describe('truncateAddress', () => {
+        it('returns empty string when address is not defined', () => {
+            expect(addressUtils.truncateAddress()).toEqual('');
+        });
+
+        it('returns input string when input is not a valid address', () => {
+            const value = '0x123';
+            expect(addressUtils.truncateAddress(value)).toEqual(value);
+        });
+
+        it('correctly truncates the address', () => {
+            const value = '0xe11bfcbdd43745d4aa6f4f18e24ad24f4623af04';
+            const expectedValue = '0xe1…af04';
+            expect(addressUtils.truncateAddress(value)).toEqual(expectedValue);
+        });
+    });
+
+    describe('getChecksum', () => {
+        it('returns the address on its checksum format', () => {
+            const value = '0xd8da6bf26964af9d7eed9e03e53415d37aa96045';
+            const checksum = '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045';
+            expect(addressUtils.getChecksum(value)).toEqual(checksum);
+        });
+
+        it('throws error when input value is not a valid address ', () => {
+            testLogger.suppressErrors();
+            const value = 'test';
+            expect(() => addressUtils.getChecksum(value)).toThrow();
+        });
+    });
+});