+++ packages/gov-ui-kit/src/modules/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './addressUtils';
+export * from './ensUtils';