+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,13 +9,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `DaoDataListItem`, `ProposalDataListItem.Structure`, and `MemberDataListItem.Structure` module components
+-   Implement `DaoDataListItem.Structure`, `ProposalDataListItem.Structure`, `MemberDataListItem.Structure` and
+    `AddressInput` module components
 -   Implement `StatePingAnimation` core component
+-   Implement `addressUtils` and `ensUtils` module utilities
+-   Implement `useDebouncedValue` core hook and `clipboardUtils` core utility
 
 ### Changed
 
 -   Update `Tag` component primary variant styling
 -   Update Eslint rules to align usage of boolean properties
+-   Update default query-client options to set a stale time greater than 0
 
 ### Fixed